Add via-stitching power circuit tests
